downstream: add background support to ivi-shell
squeshed former packages:

3fca4f1b - add background to outputs (by Janos Kovacs)
d1a2e02b - make the background color configurable (by Janos Kovacs)

the background-color entry in the ivi-shell-section defines
the color of the background. Example:
        [ivi-shell]
           ...
        background-color=1.0, 1.0, 1.0
which defines a white background.

If the background-color entry is missing or the definition is
not correct the background will be black.

Change-Id: I083471be1a3a68b56da3d6a39235ed138113b2a0
Signed-off-by: Janos Kovacs <jankovac503@gmail.com>